Add declare types in the `.udl` file
