+++ src/creds.rs
@@ -1,15 +1,10 @@
 use crate::types::*;
-use concordium_base::{
-    common::{Versioned, VERSION_0},
-    id::{
-        constants::{ArCurve, AttributeKind, IpPairing},
-        id_proof_types::{ProofVersion, StatementWithContext},
-        types::{IdentityObjectV1, IpInfo},
-    },
+use concordium_base::id::{
+    constants::{ArCurve, IpPairing, AttributeKind},
+    types::IpInfo,
 };
-use key_derivation::{CredentialContext, Net};
 use serde::{Deserialize, Serialize};
-use std::{collections::HashMap, str::FromStr};
+use std::{collections::HashMap};
 use uniffi::deps::anyhow::Context;
 use wallet_library::{
     credential::{
@@ -27,7 +22,6 @@ use wallet_library::{
         get_prf_key_aux, get_signature_blinding_randomness_aux,
         get_verifiable_credential_backup_encryption_key_aux,
         get_verifiable_credential_public_key_aux, get_verifiable_credential_signing_key_aux,
-        get_wallet,
     },
 };
 
@@ -606,213 +600,4 @@ pub fn account_credential_deployment_signed_payload(
         .map(serialize_credential_deployment_payload)
         .map_err(|e| e.to_call_failed(fn_desc.to_string()))?;
     Bytes::try_from(hex.as_str()).map_err(|e| e.to_call_failed(fn_desc.to_string()))
-}
-
-/// For the case where the verifier wants the user to show the value of an
-/// attribute and prove that it is indeed the value inside the on-chain
-/// commitment. Since the verifier does not know the attribute value before
-/// seing the proof, the value is not present here.
-///
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::RevealAttributeStatement<ArCurve, AttributeTag>`]
-#[derive(Debug, Serialize)]
-pub struct RevealAttributeStatement {
-    /// The attribute that the verifier wants the user to reveal.
-    #[serde(rename = "attributeTag")]
-    pub attribute_tag: String,
-}
-
-/// For the case where the verifier wants the user to prove that an attribute is
-/// in a range. The statement is that the attribute value lies in `[lower,
-/// upper)` in the scalar field.
-///
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::AttributeInRangeStatement<ArCurve, AttributeTag, AttributeKind>`]
-#[derive(Debug, Serialize)]
-pub struct AttributeInRangeStatement {
-    /// The attribute that the verifier wants the user to prove is in a range.
-    #[serde(rename = "attributeTag")]
-    pub attribute_tag: String,
-    /// The lower bound on the range.
-    #[serde(rename = "lower")]
-    pub lower: String,
-    #[serde(rename = "upper")]
-    /// The upper bound of the range.
-    pub upper: String,
-}
-
-/// For the case where the verifier wants the user to prove that an attribute is
-/// in a set of attributes.
-///
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::AttributeInSetStatement<ArCurve, AttributeTag, AttributeKind>`]
-#[derive(Debug, Serialize)]
-pub struct AttributeInSetStatement {
-    /// The attribute that the verifier wants the user prove lies in a set.
-    #[serde(rename = "attributeTag")]
-    pub attribute_tag: String,
-    /// The set that the attribute should lie in.
-    #[serde(rename = "set")]
-    pub set: std::collections::BTreeSet<String>,
-}
-
-/// For the case where the verifier wants the user to prove that an attribute is
-/// not in a set of attributes.
-///
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::AttributeNotInSetStatement<ArCurve, AttributeTag, AttributeKind>`]
-#[derive(Debug, Serialize)]
-pub struct AttributeNotInSetStatement {
-    /// The attribute that the verifier wants the user to prove does not lie in
-    /// a set.
-    #[serde(rename = "attributeTag")]
-    pub attribute_tag: String,
-    /// The set that the attribute should not lie in.
-    #[serde(rename = "set")]
-    pub set: std::collections::BTreeSet<String>,
-}
-
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::AtomicStatement<ArCurve, AttributeTag, AttributeKind>`]
-#[derive(Debug, Serialize)]
-#[serde(tag = "type")]
-pub enum AtomicStatement {
-    /// The atomic statement stating that an attribute should be revealed.
-    RevealAttribute {
-        #[serde(flatten)]
-        statement: RevealAttributeStatement,
-    },
-    /// The atomic statement stating that an attribute is in a range.
-    AttributeInRange {
-        #[serde(flatten)]
-        statement: AttributeInRangeStatement,
-    },
-    /// The atomic statement stating that an attribute is in a set.
-    AttributeInSet {
-        #[serde(flatten)]
-        statement: AttributeInSetStatement,
-    },
-    /// The atomic statement stating that an attribute is not in a set.
-    AttributeNotInSet {
-        #[serde(flatten)]
-        statement: AttributeNotInSetStatement,
-    },
-}
-
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::Statement<ArCurve, AttributeKind>`]
-#[derive(Debug, Serialize)]
-#[serde(transparent)]
-pub struct Statement {
-    pub statements: Vec<AtomicStatement>,
-}
-
-impl TryFrom<Statement> for concordium_base::id::id_proof_types::Statement<ArCurve, AttributeKind> {
-    type Error = serde_json::Error;
-
-    fn try_from(value: Statement) -> Result<Self, Self::Error> {
-        serde_json::to_string(&value).and_then(|s| serde_json::from_str(&s))
-    }
-}
-
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::AtomicProof<ArCurve, AttributeTag>`]
-#[derive(Debug, Deserialize)]
-#[serde(tag = "type")]
-pub enum AtomicProof {
-    /// The atomic statement stating that an attribute should be revealed.
-    RevealAttribute { attribute: String, proof: Bytes },
-    /// The atomic proof stating that an attribute is in a range.
-    AttributeInRange { proof: Bytes },
-    /// The atomic proof stating that an attribute is in a set.
-    AttributeInSet { proof: Bytes },
-    /// The atomic proof stating that an attribute is not in a set.
-    AttributeNotInSet { proof: Bytes },
-}
-
-/// Serves as a uniFFI compatible bridge to [`concordium_base::id::id_proof_types::Proof<ArCurve, AttributeKind>`]
-#[derive(Debug, Deserialize)]
-pub struct IdProof {
-    pub proofs: Vec<AtomicProof>,
-}
-
-/// Serves as a uniFFI compatible bridge to [`Versioned<concordium_base::id::id_proof_types::Proof<ArCurve, AttributeKind>>`]
-#[derive(Debug, Deserialize)]
-pub struct VersionedIdProof {
-    version: u32,
-    value: IdProof,
-}
-
-impl TryFrom<Versioned<concordium_base::id::id_proof_types::Proof<ArCurve, AttributeKind>>>
-    for VersionedIdProof
-{
-    type Error = serde_json::Error;
-
-    fn try_from(
-        value: Versioned<concordium_base::id::id_proof_types::Proof<ArCurve, AttributeKind>>,
-    ) -> Result<Self, Self::Error> {
-        serde_json::to_string(&value).and_then(|s| serde_json::from_str(&s))
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn prove_id_statement(
-    seed: Bytes,
-    net: String,
-    global_context: GlobalContext,
-    ip_info: IdentityProviderInfo,
-    identity_index: u32,
-    credential_index: u8,
-    identity_object: IdentityObject,
-    statement: Statement,
-    challenge: Bytes,
-) -> Result<VersionedIdProof, ConcordiumWalletCryptoError> {
-    let fn_name = "prove_id_statement";
-
-    let net = Net::from_str(&net)
-        .context("Failed to parse network")
-        .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-    let wallet =
-        get_wallet(hex::encode(seed), net).map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-    let ip_info = IpInfo::<IpPairing>::try_from(ip_info)
-        .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-    let identity_provider_index = ip_info.ip_identity;
-    let global_context =
-        concordium_base::id::types::GlobalContext::<ArCurve>::try_from(global_context)
-            .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-    let id_object =
-        IdentityObjectV1::<IpPairing, ArCurve, AttributeKind>::try_from(identity_object)
-            .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-
-    let credential_context = CredentialContext {
-        wallet,
-        identity_provider_index,
-        identity_index,
-        credential_index,
-    };
-    let cred_id = credential_context
-        .wallet
-        .get_prf_key(identity_provider_index.0, identity_index)
-        .context("Failed to get PRF key")
-        .and_then(|key| {
-            key.prf(&global_context.on_chain_commitment_key.g, credential_index)
-                .context("Failed to compute PRF")
-        })
-        .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-
-    let statement =
-        concordium_base::id::id_proof_types::Statement::<ArCurve, AttributeKind>::try_from(
-            statement,
-        )
-        .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-    let statement = StatementWithContext {
-        statement,
-        credential: cred_id,
-    };
-
-    let proof = statement
-        .prove(
-            ProofVersion::Version1,
-            &global_context,
-            challenge.as_ref(),
-            &id_object.alist,
-            &credential_context,
-        )
-        .context("Could not produce proof.")
-        .map_err(|e| e.to_call_failed(fn_name.to_string()))?;
-    VersionedIdProof::try_from(Versioned::new(VERSION_0, proof))
-        .map_err(|e| e.to_call_failed(fn_name.to_string()))
 }
