+++ src/lib.rs
@@ -3,6 +3,7 @@ mod encrypted_amounts;
 mod schema;
 mod transactions;
 mod types;
+mod id_proofs;
 
 // UniFFI book: https://mozilla.github.io/uniffi-rs/udl_file_spec.html
 uniffi::include_scaffolding!("lib");
@@ -12,3 +13,4 @@ pub use encrypted_amounts::*;
 pub use schema::*;
 pub use transactions::*;
 pub use types::*;
+pub use id_proofs::*;
