+++ src/lib.udl
@@ -377,6 +377,91 @@ dictionary EncryptionKeys {
     Bytes public;
 };
 
+/// For the case where the verifier wants the user to show the value of an
+/// attribute and prove that it is indeed the value inside the on-chain
+/// commitment. Since the verifier does not know the attribute value before
+/// seing the proof, the value is not present here.
+dictionary RevealAttributeStatementV1 {
+    /// The attribute that the verifier wants the user to reveal.
+    string attribute_tag;
+};
+
+/// For the case where the verifier wants the user to prove that an attribute is
+/// in a range. The statement is that the attribute value lies in `[lower,
+/// upper)` in the scalar field.
+dictionary AttributeInRangeStatementV1 {
+    /// The attribute that the verifier wants the user to prove is in a range.
+    string attribute_tag;
+    /// The lower bound on the range.
+    string lower;
+    /// The upper bound of the range.
+    string upper;
+};
+
+/// For the case where the verifier wants the user to prove that an attribute is
+/// in a set of attributes.
+dictionary AttributeInSetStatementV1 {
+    /// The attribute that the verifier wants the user prove lies in a set.
+    string attribute_tag;
+    /// The set that the attribute should lie in.
+    sequence<string> set;
+};
+
+/// For the case where the verifier wants the user to prove that an attribute is
+/// not in a set of attributes.
+dictionary AttributeNotInSetStatementV1 {
+    /// The attribute that the verifier wants the user to prove does not lie in
+    /// a set.
+    string attribute_tag;
+    /// The set that the attribute should not lie in.
+    sequence<string> set;
+};
+
+/// Statements are composed of one or more atomic statements.
+/// This type defines the different types of atomic statements.
+[Enum]
+interface AtomicStatementV1 {
+    /// The atomic statement stating that an attribute should be revealed.
+    RevealAttribute(RevealAttributeStatementV1 statement);
+    /// The atomic statement stating that an attribute is in a range.
+    AttributeInRange(AttributeInRangeStatementV1 statement);
+    /// The atomic statement stating that an attribute is in a set.
+    AttributeInSet(AttributeInSetStatementV1 statement);
+    /// The atomic statement stating that an attribute is not in a set.
+    AttributeNotInSet(AttributeNotInSetStatementV1 statement);
+};
+
+/// A statement is a list of atomic statements.
+dictionary StatementV1 {
+    sequence<AtomicStatementV1> statements;
+};
+
+/// The different types of proofs, corresponding to the statements above.
+[Enum]
+interface AtomicProofV1 {
+    /// The atomic statement stating that an attribute should be revealed.
+    RevealAttribute(string attribute, Bytes proof);
+    /// The atomic proof stating that an attribute is in a range.
+    AttributeInRange(Bytes proof);
+    /// The atomic proof stating that an attribute is in a set.
+    AttributeInSet(Bytes proof);
+    /// The atomic proof stating that an attribute is not in a set.
+    AttributeNotInSet(Bytes proof);
+};
+
+/// A proof of a statement, composed of one or more atomic proofs.
+dictionary ProofV1 {
+    sequence<AtomicProofV1> proofs;
+};
+
+/// A versioned variant of `ProofV1`
+dictionary VersionedProofV1 {
+    /// The version associated with the value
+    u32 version;
+    /// The proof
+    ProofV1 value;
+};
+
 namespace crypto {
     /// Compute the IdCredSec for the provided seed and identity indexes.
     /// Supported values for `network`: "Testnet", "Mainnet".
@@ -521,4 +606,8 @@ namespace crypto {
     /// Attempt to combine two encrypted amounts
     [Throws=ConcordiumWalletCryptoError]
     Bytes combine_encrypted_amounts(Bytes left, Bytes right);
+
+    /// Prove the supplied statement in the context of a given identity and credential
+    [Throws=ConcordiumWalletCryptoError]
+    VersionedProofV1 prove_statement_v1(Bytes seed, string net, GlobalContext global_context, IdentityProviderInfo ip_info, u32 identity_index, u8 credential_index, IdentityObject identity_object, StatementV1 statement, Bytes challenge);
 };
